script/minio_fastapi/requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
minio==7.1.15
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
